SAGE-005 Conditional render for /portfolio
Added portfolio folder to components to keep portfolio related elements
together
Added portfolio-view placeholder for viewing a portfolio
Moved create-portfolio out of page.tsx to /portfolio for condtional
render
