+++ frontend/src/components/portfolio/portfolio-view.tsx
@@ -1,0 +1,7 @@
+export default function Portfolio() {
+    return (
+        <div>
+            Portfolio placeholder
+        </div>
+    )
+}
